feat(admin): add RUM monitoring tab
