+++ apps/admin/src/pages/Monitoring.tsx
@@ -1,13 +1,13 @@
 import { useState } from "react";
 
-import Telemetry from "./Telemetry";
+import RumTab from "../features/monitoring/RumTab";
 import RateLimitTools from "./RateLimitTools";
 import CacheTools from "./CacheTools";
 import AuditLog from "./AuditLog";
 import Jobs from "./Jobs";
 
 const tabs = [
-  { id: "rum", label: "RUM", component: <Telemetry /> },
+  { id: "rum", label: "RUM", component: <RumTab /> },
   { id: "rate-limits", label: "Rate limits", component: <RateLimitTools /> },
   { id: "cache", label: "Cache", component: <CacheTools /> },
   { id: "audit-log", label: "Audit log", component: <AuditLog /> },
